+++ packages/web3-eth/test/integration/block/rpc.getBlockTransactionCount.test.ts
@@ -19,7 +19,6 @@ import { SupportedProviders, TransactionReceipt } from 'web3-types';
 import { Contract } from 'web3-eth-contract';
 import { Web3Eth } from '../../../src';
 import {
-	getSystemTestBackend,
 	getSystemTestProvider,
 	createTempAccount,
 	closeOpenConnection,
@@ -99,12 +98,8 @@ describe('rpc with block', () => {
 			}),
 		)('getBlockTransactionCount', async ({ block }) => {
 			const res = await web3Eth.getBlockTransactionCount(blockData[block]);
-			let shouldBe: number;
-			if (getSystemTestBackend() === 'ganache') {
-				shouldBe = blockData[block] === 'earliest' ? 0 : 1;
-			} else {
-				shouldBe = ['earliest', 'pending'].includes(String(blockData[block])) ? 0 : 1;
-			}
+			const shouldBe = ['earliest', 'pending'].includes(String(blockData[block])) ? 0 : 1;
+			
 			expect(Number(res)).toBe(shouldBe);
 		});
 	});