+++ packages/web3/hardhat.config.js
@@ -1,3 +1,0 @@
-// A dummy file to prevent `hardhat` from throwing
-// "HardhatError: HH1: You are not inside a Hardhat project."
-// Note: Hardhat is just used inside the tests to ensure its compatibility.