+++ packages/web3/test/e2e/e2e_utils.ts
@@ -16,24 +16,24 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 /**
- * @NOTE This Util method is kept seperate from shared system_test_utils.ts file because
+ * @NOTE This Util method is kept separate from shared system_test_utils.ts file because
  * of it's import of .secrets.json. For this method to exist in shared system_test_utils.ts
  * file, the import path would be ../.secrets.json which doesn't resolve when the file is
  * copied over to each package's test directory. Because web3 package is the only package
  * running these E2E tests that use Sepolia and Mainnet, this util exists here for now.
  */
 
-import { getSystemTestBackend } from '../shared_fixtures/system_tests_utils';
+import { getSystemTestBackend, BACKEND } from '../shared_fixtures/system_tests_utils';
 // eslint-disable-next-line import/no-relative-packages
 import secrets from '../../../../.secrets.json';
 
 export const getSystemE2ETestProvider = (): string => {
 	if (process.env.WEB3_SYSTEM_TEST_MODE === 'http') {
-		return getSystemTestBackend() === 'sepolia'
+		return getSystemTestBackend() === BACKEND.SEPOLIA
 			? process.env.INFURA_SEPOLIA_HTTP ?? secrets.SEPOLIA.HTTP
 			: process.env.INFURA_MAINNET_HTTP ?? secrets.MAINNET.HTTP;
 	}
-	return getSystemTestBackend() === 'sepolia'
+	return getSystemTestBackend() === BACKEND.SEPOLIA
 		? process.env.INFURA_SEPOLIA_WS ?? secrets.SEPOLIA.WS
 		: process.env.INFURA_MAINNET_WS ?? secrets.MAINNET.WS;
 };
@@ -42,7 +42,7 @@ export const getE2ETestAccountAddress = (): string => {
 	if (process.env.TEST_ACCOUNT_ADDRESS !== undefined) {
 		return process.env.TEST_ACCOUNT_ADDRESS;
 		// eslint-disable-next-line no-else-return
-	} else if (getSystemTestBackend() === 'sepolia' || getSystemTestBackend() === 'mainnet') {
+	} else if (getSystemTestBackend() === BACKEND.SEPOLIA || getSystemTestBackend() === BACKEND.MAINNET) {
 		return secrets[getSystemTestBackend().toUpperCase() as 'SEPOLIA' | 'MAINNET'].ACCOUNT
 			.address;
 	}
@@ -64,7 +64,7 @@ export const getAllowedSendTransaction = (): boolean => {
 
 		return Boolean(process.env.ALLOWED_SEND_TRANSACTION);
 		// eslint-disable-next-line no-else-return
-	} else if (getSystemTestBackend() === 'sepolia' || getSystemTestBackend() === 'mainnet') {
+	} else if (getSystemTestBackend() === BACKEND.SEPOLIA || getSystemTestBackend() === BACKEND.MAINNET) {
 		return secrets[getSystemTestBackend().toUpperCase() as 'SEPOLIA' | 'MAINNET']
 			.ALLOWED_SEND_TRANSACTION;
 	}
@@ -76,7 +76,7 @@ export const getE2ETestAccountPrivateKey = (): string => {
 	if (process.env.TEST_ACCOUNT_PRIVATE_KEY !== undefined) {
 		return process.env.TEST_ACCOUNT_PRIVATE_KEY;
 		// eslint-disable-next-line no-else-return
-	} else if (getSystemTestBackend() === 'sepolia' || getSystemTestBackend() === 'mainnet') {
+	} else if (getSystemTestBackend() === BACKEND.SEPOLIA || getSystemTestBackend() === BACKEND.MAINNET) {
 		return secrets[getSystemTestBackend().toUpperCase() as 'SEPOLIA' | 'MAINNET'].ACCOUNT
 			.privateKey;
 	}